+++ py_src/jupyterlab_graph_lsp/servers/__init__.py
@@ -1,2 +1,2 @@
-# Copyright (c) 2021 Dane Freeman.
+# Copyright (c) 2022 jupyterlab-graph-lsp contributors.
 # Distributed under the terms of the Modified BSD License.