+++ py_src/jupyterlab_graph_lsp/specs/graphql_language_server.py
@@ -1,9 +1,9 @@
-# Copyright (c) 2021 Dane Freeman.
+# Copyright (c) 2022 jupyterlab-graph-lsp contributors.
 # Distributed under the terms of the Modified BSD License.
 
 from jupyter_lsp.specs.utils import PythonModuleSpec
 
-URL = "https://github.com/stardog-union/stardog-language-servers/"
+URL = "https://github.com/stardog-union/stardog-language-servers"
 
 
 class GRAPHQLLanguageServer(PythonModuleSpec):
@@ -12,18 +12,18 @@ class GRAPHQLLanguageServer(PythonModuleSpec):
     python_module = "jupyterlab_graph_lsp.servers.graphql"
     key = "stardog-graphql-language-server"
     script = ["dist", "cli.js"]
-    languages = ["graphql"]
+    languages = ["graphql", "graphqls"]
     args = ["--stdio"]
     spec = dict(
         display_name="graphql-language-server",
         mime_types=["application/graphql"],
         urls=dict(
-            home=URL + "tree/master/packages/{}".format(key),
-            issues=URL + "issues",
+            home=f"{URL}/tree/master/packages/{key}",
+            issues=f"{URL}/issues",
         ),
         install=dict(
-            npm="npm install --save-dev {}".format(key),
-            yarn="yarn add --dev {}".format(key),
-            jlpm="jlpm add --dev {}".format(key),
+            npm=f"npm install --save-dev {key}",
+            yarn=f"yarn add --dev {key}",
+            jlpm=f"jlpm add --dev {key}",
         ),
     )