+++ src/index.ts
@@ -1,4 +1,5 @@
-// Copyright (c) 2021 Dane Freeman.
+// Copyright (c) 2022 jupyterlab-graph-lsp contributors.
 // Distributed under the terms of the Modified BSD License.
 
 export * from './tokens';
+export * from './icons';